Allow multiple webhooks to be created (#882)
The PR prepares gatekeeper to be able to create multiple
webhooks. Common webhook code was extracted to a separate
file, and an update to cert rotator is used that allows
to update multiple webhooks.

Tested:
Gatekeeper with the changes was run. An incoming resource
was successfuly denied request due to gatekeeper policy.
